Create the excel file with the results
