+++ README.md
@@ -45,10 +45,10 @@ python3 setup.py install
 
 #### Execute
 Run the "Search Engine Revenue" spark job with spark-submit. The 'source' and 'target' path specify the location where the source file is located and the target folder where
-the results can be stored. The 'target' folder path requires the word 'DATE' in the name, so the system can replace the word('DATE') with the current date(2019-02-07).
+the excel file will be stored.
 Here is an example of how to execute the job locally:
 ```bash
-spark-submit adobe/omniture/se_revenue/se_revenue_driver.py --source "tests/resources/data.sql" --target "tests/results/DATE_SearchKeywordPerformance.tab"
+spark-submit adobe/omniture/se_revenue/se_revenue_driver.py --source "tests/resources/data.sql" --target "tests/results/"
 ```
 The results of the run will stored in the target folder path as a CSV file . The CSV file will have the following file name prefix(and a csv extension): part-00000-*.csv
 ```bash
@@ -119,12 +119,12 @@ Note that Python3, Java 8 & Spark 2.3.3 must be installed for `pytest` to run Sp
 
 
 #### AWS-EMR (Work in Progress)
-The instructions below are to help deploy the artifact in the AWS EMR cluster. The best approach for deploying a python 
-application in AWS EMR will be to pull the artifact from an artifact repository during the cluster bootstrap. Since we 
-dont have access to an artifact repository, the artifact file and the bootstrap script file will be stored in S3. During the 
-cluster creationg, bootstrap execution will pull the artifact from S3 and install it in all the nodes on the cluster. 
- 
-Run the following command to create the project archive file in a tar.gz file. 
+The instructions below are to help deploy the artifact in the AWS EMR cluster. The best approach for deploying a python
+application in AWS EMR will be to pull the artifact from an artifact repository during the cluster bootstrap. Since we
+dont have access to an artifact repository, the artifact file and the bootstrap script file will be stored in S3. During the
+cluster creationg, bootstrap execution will pull the artifact from S3 and install it in all the nodes on the cluster.
+
+Run the following command to create the project archive file in a tar.gz file.
 ```bash
 python3 setup.py sdist
 ```
@@ -133,10 +133,8 @@ Use the aws-cli command to copy the bootstrap script and project archive file to
 aws s3 cp aws_emr/bootstrap.sh s3://adobeomniture/bootstrap.sh
 aws s3 cp dist/adobe-omniture-1.0.0.tar.gz s3://adobeomniture/adobe-omniture-1.0.0.tar.gz
 ```
-Change directory to aws_emr folder and run the aws_emr_script python script to create the cluster and submit the step: 
+Change directory to aws_emr folder and run the aws_emr_script python script to create the cluster and submit the step:
 ```bash
 cd aws_emr
 python3 aws_emr_script.py
 ```
-
-
