+++ adobe/omniture/se_revenue/se_revenue_driver.py
@@ -1,4 +1,3 @@
-import os
 from datetime import datetime
 from urllib.parse import urlsplit, parse_qs
 
@@ -11,8 +10,32 @@ from pyspark.sql.window import Window
 from adobe.omniture.se_revenue.schema import data_schema
 from adobe.omniture.utils.arg_parse import ArgParser
 from adobe.omniture.utils.csv import csv_read
-from adobe.omniture.utils.excel import create_excel_spreadsheet
 from adobe.omniture.utils.logger import Logger
+
+
+from typing import List, Tuple
+import pandas as pd
+import shutil, os
+
+def create_excel_spreadsheet(dfs: List[Tuple[str, pd.DataFrame]], target_path: str) -> None:
+    """
+    Takes a list of tuples (report names, pandas dataframes) and creates an xlsx file, where each tuple is
+    a seperate sheet.
+    :param dfs: list of tuples - names and pandas dataframes, e.g. [('adherence', adherence_df), ...]
+    :param s3_path: desired s3 location for the xlsx file
+    :return:
+    """
+    local_filename = datetime.now().strftime("%Y-%m-%d") + '_SearchKeywordPerformance.xlsx'
+    excel_writer = pd.ExcelWriter(local_filename, engine='xlsxwriter')
+    for i, tup in enumerate(dfs):
+        tup[1].to_excel(excel_writer, sheet_name=tup[0])
+    excel_writer.save()
+
+    if("s3://" in target_path):
+        os.system(f"aws s3 cp {local_filename} {target_path}{local_filename} --sse --acl bucket-owner-full-control")
+    else:
+        shutil.move(local_filename, target_path+local_filename)  # add source dir to filename
+
 
 def unpack_product_list(hit_level_df: DataFrame) -> DataFrame:
     """
