+++ requirements.txt
@@ -7,6 +7,7 @@ pyyaml==3.12
 
 #Pandas
 pandas>=0.22.0
+xlsxwriter==1.1.5
 
 # For Test
 findspark   #==1.3.0
@@ -19,4 +20,3 @@ testtools   #==2.3.0
 
 # AWS
 boto3       #==1.9.111
-
