+++ tests/test_se_revenue.py
@@ -5,7 +5,6 @@ import pytest
 from pyspark.sql import SparkSession
 
 from adobe.omniture.se_revenue import se_revenue_driver
-from adobe.omniture.utils.excel import create_excel_spreadsheet
 from adobe.omniture.utils.logger import Logger
 
 # This allows using the fixture in all tests in this module
@@ -16,7 +15,7 @@ dir_name, filename = os.path.split(os.path.abspath(__file__))
 
 # Update test config args
 source_data = os.path.join(dir_name, 'resources/data.sql')
-target_data = os.path.join(dir_name, 'results')
+target_data = os.path.join(dir_name, 'results/')
 
 job_args = {
     'source': source_data,
@@ -45,6 +44,6 @@ def test_se_revenue_pipeline(spark_session: SparkSession) -> None:
     pandas_df_list.append(("Search Engin Rev Results", raw_df.toPandas()))
 
     #Store the results locally
-    create_excel_spreadsheet(pandas_df_list, job_args['target'])
+    se_revenue_driver.create_excel_spreadsheet(pandas_df_list, job_args['target'])
 
     assert raw_df.count() == 3
